Move patch_peewee to migration.__init__
